delete unused styles, refactor author list
