+++ styles/dept.css
@@ -6,9 +6,6 @@
 }
 
 
-html.casper body {
-  font-weight: 300;
-}
 html.casper .outer {
   padding: 0 24px;
 }
@@ -36,103 +33,6 @@ html.casper .inner {
 html.casper .post-feed {
   margin: 0;
   padding: 0;
-}
-
-html.casper .post-card-image-link {
-  border-radius: 0;
-}
-
-html.casper .post-card-large .post-card-image {
-  margin-bottom: 16px;
-}
-
-@media (min-width: 795px) {
-  html.casper .post-card-large .post-card-content-link,
-  html.casper .post-card-large .post-card-meta {
-    padding-left: inherit;
-  }
-}
-
-/* Post Cards */
-html.casper .post-card-title {
-  font-size: 3.6rem;
-  line-height: 4.4rem;
-  font-weight: 300;
-}
-
-html.casper .post-card-excerpt {
-  color: #000;
-  line-height: 2.4rem;
-  font-family: 'Maison Neue';
-  font-size: 1.6rem;
-  max-width: none;
-}
-
-html.casper .post-card-meta {
-  flex-direction: column;
-}
-
-html.casper .post-card-byline-content {
-  margin-left: 0;
-  line-height: 2rem;
-  margin-bottom: 16px;
-}
-
-html.casper .post-card-byline-content a {
-  color: #000;
-  font-size: 1.4rem;
-  font-weight: 400;
-}
-html.casper .post-card-byline-date {
-  color: #000;
-  font-size: 1.4rem;
-  font-weight: 300;
-}
-html.casper .post-card-byline-date .bull {
-  margin: 0 1px;
-  opacity: 1;
-}
-
-
-/* Homepage Featured Post */
-.first-post-wrapper {
-  background: var(--onyx);
-  margin-bottom: 76px;
-  width: 100%;
-}
-
-html.casper .post-card-large .post-card-title {
-  color: #fff;
-  margin-bottom: 2.4rem;
-}
-html.casper .post-card-large .post-card-content-link
-html.casper .post-card-large .post-card-excerpt {
-  color: #fff;
-}
-
-html.casper .post-card-large .post-card-excerpt {
-  color: #fff;
-}
-
-html.casper .post-card-large {
-  padding: 52px 0;
-  margin: 0;
-}
-
-
-html.casper .post-card-large .post-card-byline-date,
-html.casper .post-card-large .post-card-byline-content a {
-  color: #fff;
-}
-
-html.casper .grid-wrapper .post-card {
-  padding-left: 0;
-  padding-right: 0;
-}
-
-html.casper .no-image .post-card-primary-tag {
-  margin-top: 0;
-  margin-bottom: 15px;
 }
 
 /* Single Post Page */
@@ -290,35 +190,6 @@ html.casper .read-next {
   background: var(--onyx);
   border-bottom: 4px solid #fff;
 }
-
-html.casper .read-next .post-card {
-  padding: 0 0 50px;
-}
-
-html.casper .up-next-header {
-  color: #fff;
-  font-size: 3.6rem;
-  font-weight: 300;
-  margin: 0;
-}
-html.casper .read-next .post-card-primary-tag {
-  color: var(--onyx);
-  opacity: 1;
-}
-
-html.casper .read-next .post-card-title {
-  opacity: 1;
-}
-
-html.casper .read-next .post-card-excerpt,
-html.casper .read-next .post-card-byline-content {
-  display: none;
-}
-
-html.casper .read-next .btn {
-  border-color: #fff;
-}
-
 
 /* Footer */
 html.casper .site-footer {
