+++ styles/screen-fixings.css
@@ -6,10 +6,6 @@ html.casper body.fade-in {
   opacity: 0;
 }
 
-html.casper .next-image-wrapper {
-  display: flex;
-}
-
 @media (min-width: 1040px) {
   html.casper .post-full-content .kg-image-card.kg-width-wide .next-image-wrapper {
     width: 1040px;
